Clean realTime data job

+++ flowgate-api/src/main/java/com/vmware/flowgate/repository/AssetRealtimeDataRepository.java
@@ -16,5 +16,5 @@ import com.vmware.flowgate.common.model.RealTimeData;
 public interface AssetRealtimeDataRepository
       extends CouchbasePagingAndSortingRepository<RealTimeData, String> {
    @Query("#{#n1ql.selectEntity} where #{#n1ql.filter} and assetID = $1 and time between $2 and $2+$3")
-   public List<RealTimeData> getDataByIDAndTimeRange(String assetID, long starttime, int duration);
+   public List<RealTimeData> getDataByIDAndTimeRange(String assetID, long starttime, long duration);
 }
